# Webpack boot script refactor