--- a/static/src/javascripts-legacy/boot.js
+++ b/static/src/javascripts-legacy/boot.js
@@ -31,9 +31,6 @@ define([
     var promiseRequire = function (moduleIds) {
         return Promise.resolve(window.require(moduleIds));
     };
-
-    console.log('Booting');
-
     var guardian = window.guardian;
     var config = guardian.config;