--- a/static/src/javascripts/boot-webpack.js
+++ b/static/src/javascripts/boot-webpack.js
@@ -1,3 +1,0 @@
-import boot from 'bootstraps/standard/main';
-
-boot();