--- a/tools/__tasks__/compile/javascript/bundle-app--webpack.js
+++ b/tools/__tasks__/compile/javascript/bundle-app--webpack.js
@@ -10,9 +10,8 @@ const { vendor, target } = require('../../config').paths;
 module.exports = {
     description: 'Create app.js for Webpack',
     task: () => Promise.all([
-        path.resolve(target, 'javascripts', 'boot-webpack.js'),
         path.resolve(vendor, 'javascripts', 'components', 'curl', 'curl-domReady.js'),
-        path.resolve(target, 'javascripts', 'boot-rjs.js'),
+        path.resolve(target, 'javascripts', 'boot-webpack.js')
     ].map(file => readFileP(file, 'utf8')))
         .then(srcs => srcs.join(';'))
         .then(allSrcs => writeFileP(path.resolve(target, 'javascripts', 'app-webpack.js'), allSrcs)),