… static/src/javascripts-legacy/boot-rjs.js