## What does this change?

This change allows us to use the same boot script for both Webpack-bundled code and r.js-bundled code. It also changes the output target to be AMD rather than the Webpack default output target.

## What is the value of this and can you measure success?

⬇️  Using the same boot script reduces the size of `app-webpack.js` from 49.9KB gzipped to 37.3KB, as dependencies are shared more efficiently. As a trade-off, `commercial-rjs.js` increases in size from 156KB to 167KB.

🛠  The build process for Webpack-bundled code had diverged confusingly from the build process for r.js-bundled code. This allows us to use a single boot script for all our bundles, so the Webpack build is easier to reason about, is closer aligned to the legacy application and should facilitate a smoother migration.

🚀  Using AMD as the output target appears to reduce the average evaluation time and the time spent recalculating layout as a result of `app-webpack.js`'s execution. Cursory performance tests showed a drop from 90ms to 64ms for the former, and 40ms to 1ms for the latter.

## Does this affect other platforms - Amp, Apps, etc?

No